end of commit

+++ php/hapus.php
@@ -1,5 +1,14 @@
 <?php
 require 'functions.php';
+
+ob_start();
+session_start();
+// Jika session login tidak ada maka kembali ke halaman login
+if (!isset($_SESSION['login'])) {
+    header('location: login.php');
+    exit();
+}
+
 $nis = $_GET['nis'];
 
 $sql = "DELETE FROM biodata_xiirpl3 WHERE nis = '$nis'";
